move to typeconv
