+++ plugins/destination/clickhouse/typeconv/arrow/unescape.go
@@ -1,4 +1,4 @@
-package client
+package arrow
 
 import (
 	"strconv"
