+++ plugins/destination/clickhouse/typeconv/typeconv.go
@@ -3,7 +3,7 @@ package typeconv
 import (
 	"github.com/apache/arrow/go/v12/arrow"
 	_arrow "github.com/cloudquery/cloudquery/plugins/destination/clickhouse/typeconv/arrow"
-	"github.com/cloudquery/cloudquery/plugins/destination/clickhouse/typeconv/ch"
+	_clickhouse "github.com/cloudquery/cloudquery/plugins/destination/clickhouse/typeconv/ch"
 )
 
 func ArrowField(name, typ string) (*arrow.Field, error) {
@@ -11,5 +11,5 @@ func ArrowField(name, typ string) (*arrow.Field, error) {
 }
 
 func ClickHouseDefinitions(fields ...arrow.Field) ([]string, error) {
-	return ch.Definitions(fields...)
+	return _clickhouse.Definitions(fields...)
 }
